fix: information form
- informationtemplate와 informationfixtemplate에서 공통적으로 사용되는
form을 모듈화 하여 재사용성과 코드 가독성을 높임
